+++ ESFramework/Foundation/NSURL+ESAdditions.h
@@ -13,11 +13,6 @@ NS_ASSUME_NONNULL_BEGIN
 @interface NSURL (ESAdditions)
 
 /**
- * Returns a Boolean value that indicates whether a given URL equals to this URL.
- */
-- (BOOL)isEqualToURL:(NSURL *)anotherURL;
-
-/**
  * Returns the query components parsed as a dictionary for this URL.
  * For URL http://foo.bar?key=value&arr[]=value&arr[]=value1 , the query component will be:
  * { key:value, arr:[value, value1] }.